+++ src/types/mypage/EmojiData.ts
@@ -1,0 +1,7 @@
+export type EmojiData = {
+  happy: number;
+  good: number;
+  soso: number;
+  bad: number;
+  tired: number;
+};
