+++ src/types/mypage/UserData.ts
@@ -1,0 +1,12 @@
+export type UserData = {
+  id: string;
+  created_at: string;
+  name: string;
+  email: string;
+  profile_image?: string;
+  gender?: string;
+  birthday?: string;
+  nickname: string;
+  main_diary?: string;
+  bloodtype?: string;
+};
